major alarm reimplementation
Reimplemented the scheduling and cancellation of alarms.
Implemented an automatic rescheduling of alarms after a reboot, app update or if the alarm is just not started at app start.
started with merging the new scheduling system with the old alarm system. Therefore the alarms currently do not work as expected