fixed duplicating journal entries and improved alarm
Fixed an issue which caused all entries in the dream journal view to be duplicated after deleting, editing or creating a dream journal entry.
Improved alarms by utilizing the AlarmManagers setAlarmClock method. This should cause the alarm to be more precise and the alarm indicator in the status bar should show up
Fixed alarm time inconsistency in alarm entries and alarm creator