implemented retrieving of notifications
It is now possible to retrieve notifications from enabled notification categories with notification messages for their selected obfuscation level. Features for pausing and disabling all notifications as well as seeing the current notification statistics are not implemented yet.
Also fixed an issue with the sizing of rod graphs which caused rod graphs not to be displayed on the statistics page.