Migrate to Kotlin and refactor code
Most of the code has been migrated to Kotlin with some Java code left which will be migrated in the future. The current state has not been thoroughly tested yet and crashes are likely to happen. Improvements include heavier use of coroutines / asynchronous tasks to improve responsiveness and performance as well as many other small improvements, refactors and general code cleanup